requirements.txt
numpy
mxnet-cu90
matplotlib
lru-dict
scipy
scikit-learn